+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # История изменений
 
+## [4.37.1](https://github.com/diplodoc-platform/components/compare/v4.37.0...v4.37.1) (2025-09-30)
+
+
+### Bug Fixes
+
+* fixed LangControl style ([#466](https://github.com/diplodoc-platform/components/issues/466)) ([af7bf9f](https://github.com/diplodoc-platform/components/commit/af7bf9fdebf9ff4bcc410297e3e8d2997758c83c))
+
+## [4.37.0](https://github.com/diplodoc-platform/components/compare/v4.36.1...v4.37.0) (2025-09-26)
+
+
+### Features
+
+* **DocPage:** added new optional prop beforeSubNavigationContent ([#458](https://github.com/diplodoc-platform/components/issues/458)) ([148576a](https://github.com/diplodoc-platform/components/commit/148576a622fa97c6b85c05c9a909ce07b55a8a8f))
+
+
+### Bug Fixes
+
+* fixed links text ([#465](https://github.com/diplodoc-platform/components/issues/465)) ([b77508f](https://github.com/diplodoc-platform/components/commit/b77508f5a3bd5b90e26ac96565b9895ab749c8ce))
+
+## [4.36.1](https://github.com/diplodoc-platform/components/compare/v4.36.0...v4.36.1) (2025-09-25)
+
+
+### Bug Fixes
+
+* fixed mobile language switching ([#462](https://github.com/diplodoc-platform/components/issues/462)) ([8b23c8b](https://github.com/diplodoc-platform/components/commit/8b23c8b307bd4a30fadd4b5d712a3b1fa8f94102))
+
+## [4.36.0](https://github.com/diplodoc-platform/components/compare/v4.35.0...v4.36.0) (2025-09-22)
+
+
+### Features
+
+* added errorTitle to ErrorPage ([#460](https://github.com/diplodoc-platform/components/issues/460)) ([5aea0e6](https://github.com/diplodoc-platform/components/commit/5aea0e61cdd196d2e7448f648ec86b011bb1b129))
+
+## [4.35.0](https://github.com/diplodoc-platform/components/compare/v4.34.2...v4.35.0) (2025-09-18)
+
+
+### Features
+
+* added availableLangs ([#454](https://github.com/diplodoc-platform/components/issues/454)) ([3f3dbb9](https://github.com/diplodoc-platform/components/commit/3f3dbb9929c68e76f3c47898a7399e263914c132))
+
 ## [4.34.2](https://github.com/diplodoc-platform/components/compare/v4.34.1...v4.34.2) (2025-09-04)